Refactor render-engine error plumbing and consolidate FileManager helpers
- FileOutcome.result: Result<Void, RunError> -> (any Error)? (nil == success)
- Merge all SyntaxKit FileManager extensions into FileManager+Execution.swift
  (isLibDir, libStamp, collectInputs, writeOutput); drop the inline extension
  from Runner+Directory.swift (also clears its file_types_order warnings)
- runSwift backend returns SwiftRunOutcome; move .completed wrapping out of Runner
- collectInputs throws a dedicated CollectInputsError (.cliError /
  .resourceValuesFailure), mapped to RunError.unexpected by renderDirectory
