--- Sources/SyntaxKit/Execution/FileManager+LibStamp.swift
@@ -1,43 +1,0 @@
-//
-//  FileManager+LibStamp.swift
-//  SyntaxKit
-//
-//  Created by Leo Dion.
-//  Copyright © 2026 BrightDigit.
-//
-//  Permission is hereby granted, free of charge, to any person
-//  obtaining a copy of this software and associated documentation
-//  files (the “Software”), to deal in the Software without
-//  restriction, including without limitation the rights to use,
-//  copy, modify, merge, publish, distribute, sublicense, and/or
-//  sell copies of the Software, and to permit persons to whom the
-//  Software is furnished to do so, subject to the following
-//  conditions:
-//
-//  The above copyright notice and this permission notice shall be
-//  included in all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
-//  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-//  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-//  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-//  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-//  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-//  OTHER DEALINGS IN THE SOFTWARE.
-//
-
-import Foundation
-
-extension FileManager {
-  /// `<size>/<mtime>` fingerprint of `libSyntaxKit.{dylib,so}` under
-  /// `libPath`, or nil if unreadable. Catches in-place rebuilds without a
-  /// version bump.
-  internal func libStamp(libPath: String) -> String? {
-    let dylib = "\(libPath)/\("SyntaxKit".dylibFilename)"
-    guard let attrs = try? attributesOfItem(atPath: dylib) else { return nil }
-    let size = (attrs[.size] as? NSNumber)?.intValue ?? 0
-    let mtime = (attrs[.modificationDate] as? Date)?.timeIntervalSince1970 ?? 0
-    return "\(size)/\(Int(mtime))"
-  }
-}
